+++ lib/src/extension.dart
@@ -1,3 +1,5 @@
 export 'package:token_parser/src/extensions/iterable.dart';
 export 'package:token_parser/src/extensions/pattern.dart';
+export 'package:token_parser/src/extensions/lexeme.dart';
 export 'package:token_parser/src/extensions/top_level.dart';
+export 'package:regex_range/regex_range.dart' show range;